User can only edit at initial sim state / Stop button

+++ src/bridge/controllers/BodyController.js
@@ -20,11 +20,12 @@ angular.module('bridge.controllers')
     $scope.timestepUnits = 'seconds';
     $scope.simulator = simulator;
     $scope.Units = Units;
+    $scope.eventPump = eventPump;
     $scope.User = User;
 
     this.getRunningState = function() {
-      return !eventPump.paused;
-    }
+      return !eventPump.paused || !simulator.isEditable();
+    };
 
     this.selectCenterBody = function(){
         $scope.simulator.orbitTracker.setCenterBody(
@@ -49,7 +50,7 @@ angular.module('bridge.controllers')
     };
 
     this.updateBody = function(body){
-      if(User.current) {
+      if(User.current && simulator.isEditable()) {
         simulator.updateBody(body.id, body);
         eventPump.step(false, true);
       }
@@ -66,7 +67,7 @@ angular.module('bridge.controllers')
     };
 
     this.remove = function(id) {
-      if(User.current){
+      if(User.current && simulator.isEditable()){
         $('#' + id).attr('r', 0);
         simulator.deleteBody(id);
         eventPump.step(false,true);
